backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
django-filter==25.1
openai>=1.35.0
